@@ -1,17 +1,16 @@
 package io.mosip.kernel.saltgenerator.step;
 
-import java.util.List;
 import java.util.stream.Collectors;
+import java.util.stream.StreamSupport;
 
 import javax.transaction.Transactional;
 
+import org.springframework.batch.item.Chunk;
 import org.springframework.batch.item.ItemWriter;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
 import io.mosip.kernel.core.logger.spi.Logger;
-import io.mosip.kernel.core.saltgenerator.constant.SaltGeneratorErrorConstants;
-import io.mosip.kernel.core.saltgenerator.exception.SaltGeneratorException;
 import io.mosip.kernel.saltgenerator.entity.SaltEntity;
 import io.mosip.kernel.saltgenerator.logger.SaltGeneratorLogger;
 import io.mosip.kernel.saltgenerator.repository.SaltRepository;
@@ -35,13 +34,14 @@ public class SaltWriter implements ItemWriter<SaltEntity> {
 	 */
 	@Override
 	@Transactional
-	public void write(List<? extends SaltEntity> entities) throws Exception {
-		if (repo.countByIdIn(entities.parallelStream().map(SaltEntity::getId).collect(Collectors.toList())) == 0l) {
+	public void write(Chunk<? extends SaltEntity> entities) throws Exception {
+		if (repo.countByIdIn(StreamSupport.stream(entities.spliterator(), true).map(SaltEntity::getId).collect(Collectors.toList())) == 0l) {
 			repo.saveAll(entities);
 			mosipLogger.debug("SALT_GENERATOR", "SaltWriter", "Entities written", String.valueOf(entities.size()));
 		} else {
 			mosipLogger.error("SALT_GENERATOR", "SaltWriter", "write", "Records already exists");			
 		}
 	}
+
 
 }
