@@ -3,14 +3,13 @@ package io.mosip.kernel.saltgenerator.step.test;
 import static org.junit.Assert.assertEquals;
 import static org.mockito.Mockito.when;
 
-import java.util.Collections;
-
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.junit.MockitoJUnitRunner;
+import org.springframework.batch.item.Chunk;
 
 import io.mosip.kernel.core.saltgenerator.constant.SaltGeneratorErrorConstants;
 import io.mosip.kernel.core.saltgenerator.exception.SaltGeneratorException;
@@ -32,7 +31,7 @@ public class SaltWriterTest {
 		SaltEntity entity = new SaltEntity();
 		entity.setId(0l);
 		when(repo.countByIdIn(Mockito.any())).thenReturn(0l);
-		writer.write(Collections.singletonList(entity));
+		writer.write(Chunk.of(entity));
 	}
 
 	@Test
@@ -41,7 +40,7 @@ public class SaltWriterTest {
 			SaltEntity entity = new SaltEntity();
 			entity.setId(0l);
 			when(repo.countByIdIn(Mockito.any())).thenReturn(1l);
-			writer.write(Collections.singletonList(entity));
+			writer.write(Chunk.of(entity));
 		} catch (SaltGeneratorException e) {
 			assertEquals(SaltGeneratorErrorConstants.RECORD_EXISTS.getErrorCode(), e.getErrorCode());
 			assertEquals(SaltGeneratorErrorConstants.RECORD_EXISTS.getErrorMessage(), e.getErrorText());
